refactor(web): move VercelTokenDialog to shared components, fix assistantId guard
Move VercelTokenDialog from domains/intelligence/ to src/components/
since it's now consumed by both the chat and intelligence domains.
This removes the new cross-domain allowlist entry (convention says
the list should only shrink, not grow).

Also gate the dialog render on assistantId being present instead of
falling back to an empty string — avoids a silent footgun if the
component ever makes API calls with an empty ID.
